[i2c,dv] Fixup the i2c_host_mode_toggle_vseq
We don't model this sort of mid-transfer disabling of the Controller-Module at
all in the DV environment, so the current approach is just to disable the
scoreboard and reset any internal state that would cause the testing of the
recovery stimulus to fail.

In the future, it would be great to model this properly, and hence leave the
scoreboard/ref_model enabled the entire time. Since this sort of mid-transfer
disable of the block is a valid use-case for ending a transfer early, and is
described in the theory_of_operations/programmers_guide, we should endeavour to
model it.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>