[rv_dm,dv] Extend rv_dm_ndmreset_req_vseq to check separate rst pins
This doesn't really change the behaviour of the sequence except that
we now do a slightly better job of representing the rest of the
system, by reporting a reset through the clk_rst_lc_ni pin.

However, we slightly tighten the test by making sure that rv_dm's
tracking of the pending reset matches what we expect. We also check
that the anyhavereset and allhavereset bits go high in the way we'd
expect.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>