[hmac,dv] remove invalid assertion
This assertion was checking that a `process` or `stop` command happened
only after a `start` or `continue` command had been issued. However, it
was doing so by checking if the `process` or `stop` signals being high
overlapped with the `initiated` signal being high (which simply latches
whether a `start` or `continue` had been seen in the past). However,
multiple `process` commands may be issued by SW before a `stop` or `done`,
and the hardware does support this. Any ROM E2E test that enabled SPX+
signature verification does indeed trigger multiple `process` operations.
This was the root cause behind #23954.

Since, what this assertion was checking was too restrictive, we opt for
simply removing it.

This fixes #23954.

Signed-off-by: Tim Trippel <ttrippel@google.com>