[rv_dm,dv] Update RAL on observed DMI reads
We still don't do predictions (and probably should do), but this
change uses the predict method to update the RAL model of a CSR when
we see a read of its value.

The existing code was a bit silly: this was just a few lines too late
and was outside of the loop.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>