[jtag,dv] Fix monitor tracking of trst_n
The existing logic didn't actually work if we saw a reset that went
away before the next clock edge. If that happened, the monitor would
completely miss the reset because we were in the middle of wait_tck().

With this tweaked version, we wait until the first of "reset" and "tck
edge", then spot a reset has happened because trst_n is low.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>