[silicon_creator,hmac] ensure `process` cmd is only invoked once
This updates the HMAC silicon_creator driver to separate out the
`hmac_sha256_process()` call from the `hmac_sha256_final...()` calls to
avoid sending multiple process commands to the hardware, and thus
triggering an SVA in sim (see #23954). While the consensus in #24022 was
that the hardware can handle multiple `process` commands, this path is
less tested. Hence, we make this update to minimize risk in the ROM code
that drives the HMAC SHA2 engine.

This updates the ROM, and fixes #23954. A new rom release candidate will
tagged after this merges.

Signed-off-by: Tim Trippel <ttrippel@google.com>