[rv_dm,dv] Add late_debug_enable_regwen to scoreboard list of regs
This doesn't actually add any testing, but it avoids a quick error
path when a stress test tries to write the late_debug_enable_regwen
register.

Eventually, we should probably check that the register *does*
something useful. This change is much simpler than that: it just stops
the random write from causing the test to fail for a silly reason.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>