[dv] Tidy up how reset lengths get calculated
The dv_base_vseq.apply_resets_concurrently task wants to wait at least
two clock cycles. It also takes a reset_duration_ps argument.

Before this change, we calculated the slowest clock period by taking
the maximum of the clock periods and the reset_duration_ps argument.
We then scaled by a number in the range 2-10 to ensure that we waited
at least two cycles.

After this change, we calculate the slowest clock period and scale
that up, but then take the maximum of the result and the
reset_duration_ps argument. This avoids scaling up the thing that's
called a reset duration by treating it as a clock period.

It also shortens some resets in rv_dm by a factor of up to 10!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>