[dv] Track trst_n reset of IR in jtag_dmi_monitor.sv
This means that we monitor things correctly if we see a DR transaction
immediately after a reset.

Honestly, this is a little silly: it would probably be better if the
jtag monitor sent the IR for each transaction (so that we didn't have
to implement IR tracking on everything downstream). But that's a
rather bigger change.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>