[aes/dv] Check DUT status in aes_reseed_vseq.sv using a backdoor access
Depending on the latency of the TL-UL and the EDN agents, the first read
to the status register might only return after the reseeding has
finished already, which then causes a timeout as no reseed operation can
be observed. Using a backdoor access to check the status reliably solves
the problem.

This resolves lowRISC/OpenTitan#24052.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>