[dv] Add documentation to tl_write_mem_less_than_word task
Also slightly change the arguments so that we pass the ranges of
memory relative the base of the block as an argument (and avoid the
"updated" prefix, which I found very confusing!)

Also, do the pruning that finds memories that don't support partial
writes before we start the loop. I think this makes the code a bit
easier to follow.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>