[dv] Add documentation to tl_read_wo_mem_err
Also move the filtering that finds the write-only memories so that it
runs outside of the loop. This should be slightly clearer to read.
Finally, change things so we pass the register block and list of
relative memory address ranges as arguments, which avoids so much
being tied in with class variables.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>