[sw,lib] Add compile-time check for log argument size
When an argument (to be formatted via format specifiers in a given string) is passed to the
`LOG` function either directly or indirectly, this value is listed as a variable argument,
and then read from the va_list. In the case of DV logging via the backdoor, this is always
read as a `uint32_t`. In the case of core logging (on Ibex), this is either `uint32_t`,
`size_t`, `uintptr_t` or `status_t` (a struct storing an `int32_t`).

As a result, when a 64 bit integer is passed into some debug logging functionality without
appropriate explicit casting, no implicit casting is performed as may be expected by a
programmer, and instead the 64 bit integer argument is read as a 32 bit variable argument.

As per N1570 7.16.1.1 paragraph 2: "... if type is not compatible with the type of the
actual next argument (as promoted according to the default argument promotions), the
behaviour is undefined, except for the following cases:
    - one type is a singed integer type, the other type is the corresponding unsigned
    integer type, and the value is representable in both types.
    - one type is pointer to void and the other is a pointer to a character type."

Because there is no implicit cast, any such invocations of debug printing functionality
without casting 64 bit arguments are undefined behaviour, and may cause unreliable and
unexpected behaviour in logging. This can compound and mislead separate debugging efforts,
and so should not be allowed to occur.

This commit introduces additional macro functionality in order to check the type of each
of these variable arguments at compile time. As a result, any code that provides an
incorrect argument to e.g. the `LOG_INFO` function will result in the code not compiling
with an appropriate error message pointing out the line and a suggestion for how to fix
the problem. As this is all done via `_Generic` and `static_assert`, no additional
runtime cost is added to perform these checks.

As part of this functionality, this commit expands the base library of macros to include
generic macros for invoking a macro per each argument in a variable list, as well as
checking the size of a given argument.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit 17787ea12521ba5d1f99de7cd9fd42b095fd93e2)