[aes/dv] Fix corner cases in aes_reseed_vseq.sv
This commit fixes a couple of rare corner cases in aes_reseed_vseq.sv.
In particular, it was previously possible that due to varying latency
of the TL-UL agent, the response of the operation triggering a PRNG
reseed (e.g. writing the trigger register, writing the last key
register) finished only after the reseeding of the clearing PRNG was
done.

To avoid this issue, operation triggering the reseed and the checking
of the reseed can be done in two separate threads. This works perfectly
except for the case where we want to check that no reseed operation is
happening. To this end, the checking function is split into two
separate functions.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>