[entropy_src/dv] Test entropy drops for max rate test
This commit chnages the environment and the rng vseq for the
RNG max rate test. When running at max rate the DV used to limit
the delay of the halt interface to avoid entropy drops. Since now
entropy drops are modelled, this doesn't need to be the case anymore.
Furthermore this commit disables some assertions that expect no drops.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>