[sw,lib] Fix debug log argument errors in tests
Fix all compilation errors in bazel targets that occur as a result
of the compile-time logging argument checking that ensures that
uint64_t values are not passed to `LOG` or equivalent functions
(any functions like `CHECK` or `LOG_INFO` that call LOG).

This can be verified by running

```
./bazelisk.sh build -t- --keep_going --define bitstream=skip
	//util/... //sw/...
```

and then grepping the output for the specific error (for 'must be
no wider than 32 bits').

All fixes simply involve performing an appropriate cast from a 64
bit integer type to a relevant 32 bit integer type. It is assumed
based on the context of these variables that all values are safe
to cast without any loss of information; no existing formatting
utilities exist to print integers larger than 32 bits.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>