[jtag,dv] Get rid of exit_to_rti_dr_past
I'm pretty certain that the author wasn't quite modelling the current
TAP FSM state properly and the naming is rather confusing.

I think the name is probably backwards and the variable being set
means that we think we're in the Test-Logic-Reset state instead of
Run-Test/Idle. But it isn't actually always correct (and the name is
confusing). Drop it. If necessary, we can follow up by tracking things
more explictly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>