[otbn,dv] Correct timing for when INSN_CNT gets zeroed in model
This was a cycle too late if we got an escalation request when were
were in the middle of a secure wipe. The reason is that we checked
lock_after_wipe in the _delayed_insn_cnt_zero function just before
setting the flag.

The fix is to set the flag *before* checking it...

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>