[dv,usbdev] Fix usbdev_pkt_buffer sequence
The behavior of the DUT with packets that do not fully
fit within the packet buffer when the CRC16 bytes are
considered is very subtle.
Unfortunately it does have memory from one packet to another
within the internal flops, so modify the BFM to mimic this
for now. Perhaps in time the DUT should be modified to reset
the flops; since USBDEV is not considered important from a
security perspective this is a DV issue only.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>