[otp,doc] add prodc OTP image with SPX+ enabled
This adds a personalized prodc OTP image with SPX+ sigverify enabled for
the purposes of splicing the latest cached bitstreams with this OTP
image for validating owner firmware payloads.

Additionally this adds instructions to the website to explain how to use
the `universal_splice` Bazel target to splice custom bitstreams.

Signed-off-by: Tim Trippel <ttrippel@google.com>