[rom_ctrl,dv] Chop out most of rom_ctrl_env_cfg.initialize
Most of this can be done at object creation. We only really need to
override initialize because we want to follow it up by updating
max_outstanding_req on some TL interfaces.

I just checked and creation of those interfaces needs to be in
initalize (rather than the constructor) in the base class because
dv_base_env_cfg needs to know the CSR base address in order to create
the ral models that feed into TL interface creation.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>