[ci] Update quick-lint & CI scripts after site move
After the building of the site / landing page was moved to a separate
repository, there is still an existing reference to the old
`./ci/scripts/build-site.sh` script in `./ci/jobs/quick-lint.sh`.

This commit also excludes the `build-site` directory from the CI
executable file and shell checks, as current building of the docs
via mdbook creates many executable files and shell scripts which do not
pass linting standards (but are temporary, and local).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>