[rv_dm,dv] Massively simplify rv_dm_cmderr_exception_vseq
This gets rid of a magic hex constant (when I "decompiled it", this
turned out to be an abstract command to write to x0) and replaces it
with another arbitrary command (reading x0).

It also reworks the sequence by adding comments and making it a bit
clearer what's going on. We also remove the repetition of the test:
there's no real reason to believe that running it 10 times will make
any difference...

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>