[rv_dm,dv] Expand the implementation of mem_tl_access_resuming
This expands the code a bit to make it clearer what's going on. It
also drops the loop: that doesn't seem to make all that much
sense (and nor does a randomised resume request).

Honestly, this is really a cut-down version of
rv_dm_halt_resume_whereto_vseq. That sequence tries to request an
abstract command and then checks the whereto register contains the
start of that command.

It probably makes sense to keep this (more simple) test as well,
partly because the code is dramatically easier to understand.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>