[rv_dm,dv] Expand the rv_dm_progbuf_read_write_execute test
Not much of a functional change, except that I've made it so that now
we write all the multiregisters before reading them all back (to make
sure that they aren't all backed by a single register).

I've also added comments to make it as clear as possible what's going
on.

This is a cherry pick of commit 5d7217c7915df7edccc2789c5748f15de1f0df42
to branch earlgrey_1.0.0.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>