[dv,flash_ctrl] Fix word sweep tests
These tests rely on read_flash and prog_flash to increment the addresses
in their flash_op_t argument precisely by the words accessed, but it is
better to have the caller take care of it. The previous behavior caused
some failures in flash_ctrl_write_word_sweep.

This is a cherry pick of commit 16cdd48323236e64bbea0fa759149b1f199ae48f
to branch earlgrey_1.0.0.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>