[rv_dm,dv] Add manual exclusion for WhereToAddr cond coverage
There's a detailed comment about why this should be excluded above the
exclusion conditions.

Note that this is probably something that could be excluded by UNR as
well. But I spent half an hour figuring out what was going on, so
maybe it makes sense to waive it explicitly.

For a later me: To generate the syntax easily, use the following
steps.

1. Run the tests in question with coverage enabled

2. Call urg manually and get it to dump a dummy exclusion file. I
   used:

      urg -full64 \
          -dir scratch/work/rv_dm-sim-vcs/coverage/default.vdb \
          -dump full_exclusions

   (here, the path comes from dvsim. "work" is the name of the branch
   that I ran)

   You end with files called fullexclude* in the current directory.

3. Now cut the relevant lines from the relevant file and add a
   detailed explanation above them. Note: there's a reasonably helpful
   syntax description in a big comment at the top of the file.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>