[rv_dm,dv] Fix sequence in rv_dm_halt_resume_whereto_vseq
The important "correctness" change here is that the debugger now
writes 1 a last time to resumereq to clear the request again.

While I'm here, I've also changed the code so that it only runs
through once: if we want to do chaining, we should probably use the
stress sequence.

I've also fixed some comments: I originally wrote "debugger" to mean
the debug module. This is wrong! The debugger should be the thing that
is asking the debug module to do stuff. Fix that.

This should also add coverage for the condition on line 168 of
dm_mem.sv. Before this change, we were never seeing a write to
resumereq when we had the resuming flag and were in the Idle state.

Note that this doesn't actually hit one of the condition coverage
items from the if statement: when resumereq is true and resuming_q is
false. I think this is actually impossible (because resuming_q gets
cleared by clear_resumeack_i on the same cycle as resumereq gets set).
Assuming I've got that right, this should be picked up by a UNR run.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>