[bazel] Remove sim envs for usbdev_aon_wake_reset_test
Removes the `sim_dv` and `sim_verilator` execution environments from
the `usbdev_aon_wake_reset_test` bazel test target. The source file
for the test that is being run already immediately performs a check
to ensure that the device type is not DV or Verilator, and fails if
it is. This lack of support for sim environments is now made explicit
in the Bazel parameters so that this test is not accidentally run.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit 5bc109756465b9858dc30784ffc0d61928257241)