[dv,usbdev] Sequence to disrupt tx/rx mid-transaction
Disrupt transmission or reception using an immediate IP block
reset or USB Bus Reset to ensure that the device is not left
with persistent state and may resume transmission and reception
subsequently.

This is a cherry pick of commit 803c6906574d02964e7884c0c9c9cb9f97d2cc23
to branch earlgrey_1.0.0.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>