[verilator,spi] Reduce SPI clock frequency on Verilator tests
Fixes the `spi_host_irq_test` test target which was failing on Verilator
due to an incorrect SPI config. This is because the SPI clock frequency
was configured to 1 MHz, whereas Verilator's HiSpeed Peripheral
frequency is set to just 500 KHz. As such, `dif_spi_host_configure`
would calculate a divider of all 1s, guaranteeing a bad argument (so
long as `SPI_HOST_CONFIGOPTS_CLKDIV_0_MASK` is not all 1s). This commit
handles this case by checking the device type and conditionally reducing
the SPI Clock speed tenfold to 100 kHz when running in Verilator.

This change is also necessary for `spi_host_smoketest`, but that test
is broken besides this due to requiring a flash connected to the SPI
bus, so that test has been changed to remove the verilator environment
from the BUILD target accordingly.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>