[bazel] Mark broken/manual verilator target environments
This commit addresses three tests:
- For `spi_host_winbond_flash_test`, the Verilator execution environment
has been removed. This is because the test requires an SPI flash on the
bus, which is not available in the Verilator simulation environment.
- For `rv_core_ibex_mem_test_prod`, the Verilator execution environment
has been removed. This is because the `rv_core_ibex` test is marked as
being incompatible with test_rom, which Verilator uses, and as such the
Verilator execution environment should not be supported here.
- For `pwrmgr_all_reset_reqs_test`, the Verilator execution environment
is marked as broken alongside the existing broken cw310_sival_rom_ext
environment, because the test is providing the same error for both
cases. When that is fixed, it should also be tested to see if the
verilator environment is still broken or can be re-supported as well.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit ffd7bb274b6cdeb11357abbfb988cd63cf0989a7)