[bazel] Mark broken/manual verilator target environments
This commit addresses three tests:
- For `spi_host_winbond_flash_test`, the verilator execution env
is now marked as "manual", as the test requires access to winbond
flash which is not currently available in CI. As such, it should not
be run except manually.
- For `rv_core_ibex_mem_test_prod`, the Verilator execution environment
has been removed. This is because the `rv_core_ibex` test is marked as
being incompatible with test_rom, which Verilator uses, and as such the
Verilator execution environment should not be supported here.
- For `pwrmgr_all_reset_reqs_test`, the Verilator execution environment
is marked as broken alongside the existing broken cw310_sival_rom_ext
environment, because the test is providing the same error for both
cases. When that is fixed, it should also be tested to see if the
verilator environment is still broken or can be re-supported as well.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>