[verilator,tests] Mark timed-out verilator targets as manual
All of these Verilator test targets were run on public CI with timeouts
of 40 minutes to an hour.

```
./bazelisk.sh test --test_timeout=2400,2400,3600,-1 <TEST_NAME>
```

Of the 390 test targets, all 67 that timed out in this time are judged
to be too long to run in CI. This commit marks these verilator targets
as "manual" so that they are not run in CI, until further testing proves
whether they actually succeed (and run over time) or are broken and
should be marked as such.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit 1dd730fc94e176172728e34f4c61a163f74b1285)