[dv,usbdev] Improve testing of suspend transitions
Some of the link transitions were previously not observed because
(i) the driver/sequences typically do not model the USB by
periodically sending the SOF packets when the bus is Idle.
(ii) the Powered link state is transient and difficult to observe
because the DP/DN both low state before VBUS connection acts as
a Bus Reset. It therefore tends to transition quickly to ActiveNoSOF.
Update the BFM to reflect the Powered -> PoweredSuspended transition.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>