[dv,usbdev] Handle DUT reset in scoreboard/model
Modify the scoreboard and BFM to be aware of DUT reset occurring
during a sequence.
Ignore monitor traffic that is received whilst the DUT is under
reset. Discard any outdated register field predictions, reinstating
the initial values as predictions of the post-reset state of the DUT.

This is a cherry pick of commit 3993353cc572a322461b657a731b8b404cb5d452
to branch earlgrey_1.0.0.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>