[provisioning] do not generate RMA unlock token on the device
Generating the RMA unlock token on the device and then transferring it
to the host for future registering is pretty cumbersome: since the
token must not be sent over the host connection in clear a whole ECDH
key exchange and generating of an AES key on the target are required.

It is much cleaner to have the token generated on the host and have
just the hash sent to the device. No need to encrypt the hash, fewer
messages sent back and forth.

This also removes a lot of complexity from the device code.

Tested by verifying that both personalize_functest and
ft_provision_fpga test pass on the FPGA setup.

Signed-off-by: Vadim Bendebury <vbendeb@google.com>
(cherry picked from commit 10e8acafb8e5ba07bb169d7d1a2095a4d0d7ee43)