[csrng/dv] Properly trigger the status error responses
This commit changes the interrupt vseq to properly trigger
the command error status responses instead of forcing the signals.

This commit also changes the coverage for the command status response,
since it was outdated and causing failures whenever coverage was
collected.

This is a cherry pick of commit 65e5131f07fe13a939a4bb72f65631d0766b2660
to branch earlgrey_1.0.0.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>