[dv,usbdev] Optional reporting of bit stuffing violations
Allow reporting of bit stuffing violations to be disabled for
the `usbdev_tx_rx_disruption` sequence which aims to disrupt
USB traffic and DUT operation mid-transaction.
In this extreme case it can be rather impractical and require
detailed internal knowledge of the DUT to predict reliably
whether a report of a bit stuffing violation will occur.

This is a cherry pick of commit a291f72f771b2aab0a9d98242109ac877e5a88cf
to branch earlgrey_1.0.0.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>