[dv,usbdev] Optional reporting of bit stuffing violations
Allow reporting of bit stuffing violations to be disabled for
the `usbdev_tx_rx_disruption` sequence which aims to disrupt
USB traffic and DUT operation mid-transaction.
In this extreme case it can be rather impractical and require
detailed internal knowledge of the DUT to predict reliably
whether a report of a bit stuffing violation will occur.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>