[kmac/dv] Extend timeout values for long running tests
As of lowRISC/OpenTitan#23942, the FIFO empty status-type interrupt is
predicted and verified which unfortunately negatively impacts the
simulation performance in some cases. Most tests are not notably
impacted but for some rather long running tests, the maximum observed
job runtime increased by up to 50%. As a result, we started seeing
timeouts for some of these tests in the nightly regressions after
merging the FIFO empty interrupt verification PR.

Therefore, this commit increases the timeout for those long running
tests where we observed timeouts.

This resolves lowRISC/OpenTitan#24225.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>