[dv,usbdev] Fix failure in `usbdev_bitstuff_err` sequence
The sequence `usbdev_bitstuff_err` could inject bit stuffing
violations from which the current RTL does not properly recover
and it instead tries to interpret the remainder of the OUT token
packet as a separate packet, leading to a `rx_pid_err` assertion
too.
Instead, ensure that only known, recoverable bit stuffing violations
are exercised with the current RTL, pending a more robust recovery
mechanism in a later revision.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>