[rom_ext] Set the boot slot on every boot
The function `boot_log_check_or_init` is meant to initialize the
`boot_log` structure if it hasn't already been initialized.  This is
meant to initialize the boot_log if the ROM doesn't (it doesn't on the
earlgrey_es chip), however, it will also skip initialization if the
boot_log is already initialized from a prior boot.

Initialize only portions of the boot_log that will be constant across
boots.  Defer setting the `rom_ext_slot` to `rom_ext_start` and set it
to the current boot slot on every boot.

Fixes: #23869

Signed-off-by: Chris Frantz <cfrantz@google.com>