[dif, sensor_ctrl] Fix incorrect AST event validity range check
The `sensor_ctrl` DIF previously used the following range check to
determine if a supplied alert event is valid:
    event_idx > SENSOR_CTRL_PARAM_NUM_ALERT_EVENTS
This commit changes this to be <= instead (correctly excluding the
end value, as only 0 to NUM_ALERT_EVENTS - 1 are valid alerts).

This also makes a small stylistic change to remove the negation from
the function to make the code using it more readable.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>