[dv,usbdev] Shut down properly after sequence
Disconnect the DUT and revert the configuration changes after
the `usbdev_iso_retraction` sequence because otherwise it has
the potential to interfere with the attempt to clear interrupts
in cip_base_vseq (`host_lost` may become asserted at just the
wrong moment).

This is a cherry pick of commit 588ced176b3234a4399d16831113333272ee1a32
to branch earlgrey_1.0.0.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>