[rv_dm,dv] Remove bogus TODO note from me
This demonstrates nicely that I didn't properly understand JTAG when I
wrote it. I understand a bit more now and the structure with TDI
flowing into a shift register before flowing back out of TDO is
literally given in the spec. Oops.

This is a cherry pick of commit 4ec8b62ed084abba9f61051645fae87f116ceaf3
to branch earlgrey_1.0.0.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>