[dv,usbdev] Delay the response to IN token packets
Introduce a simple timing model and short delay in the BFM
before responding to IN token packets; this is neccessary
because the CSRs may be modified between the conclusion of
IN packet signaling and the point of Data/Status signaling
in response, leading to behavior that differs between the
model and the DUT.
`usbdev_iso_retraction` was particularly affected.

This is a cherry pick of commit 3707c48f56837b111c1a11afe1afe460ddee2a97
to branch earlgrey_1.0.0.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>