[rv_dm,rtl] Move an enum value into a package
This has no effect, but makes the number of copies of the debug enable
value visible from outside of rv_dm.sv. It turns out that even hacking
things in the DV code to pass the value 5 causes (rather cryptic)
compile time errors from VCS.

Note that it looks like the value of RvDmLcEnSize might have changed,
but we're changing from $clog2(6) to $clog2(5) and the two values are
equal.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>