[manuf] add post-cert-endorsment perso extension hook
Originally, there was only a single personalization extension hook that
enabled SKU owners to generate additional TBS certificates to be
endorsed by the host. This hook function was executed before the TBS
certificate data payload was sent to the host. This change adds a second
extension function to enable SKU owners to provision additional data
into flash *after* the certificate data payload has been sent from the
host back to the device, in case there are additional host-generated
certificates or other metadata that needs to be injected from the host
to the device.

Signed-off-by: Tim Trippel <ttrippel@google.com>