[manuf] add host FT lib extension hook
This adds a provisioning extension hook to the host side FT provisioning
test harness that enables SKU owners to inject additional assets into
the device that originate on the host side. This expands on the existing
FT provisioning test harness that only enabled endorsing certificates the
device had originally sent the host.

Signed-off-by: Tim Trippel <ttrippel@google.com>